+++ src/consumer/mod.rs
@@ -12,7 +12,7 @@ use rdsys::types::*;
 use client::{Context, NativeClient};
 use error::KafkaResult;
 use groups::GroupList;
-use message::Message;
+use message::BorrowedMessage;
 use metadata::Metadata;
 use util::cstr_to_owned;
 
@@ -132,13 +132,13 @@ pub trait Consumer<C: ConsumerContext> {
     /// Commit a specific message. If mode is set to CommitMode::Sync,
     /// the call will block until the message has been successfully
     /// committed.
-    fn commit_message(&self, message: &Message, mode: CommitMode) -> KafkaResult<()> {
+    fn commit_message(&self, message: &BorrowedMessage, mode: CommitMode) -> KafkaResult<()> {
         self.get_base_consumer().commit_message(message, mode)
     }
 
     /// Store offset for this message to be used on the next (auto)commit.
     /// When using this `enable.auto.offset.store` should be set to `false` in the config.
-    fn store_offset(&self, message: &Message) -> KafkaResult<()> {
+    fn store_offset(&self, message: &BorrowedMessage) -> KafkaResult<()> {
         self.get_base_consumer().store_offset(message)
     }